+++ src/panel/add_card/addActions.js
@@ -1,6 +1,6 @@
 import {action} from "mobx";
-import MainStore from "../MainStore";
-import OptionStore from "../OptionStore";
+import MainStore from "../common/MainStore";
+import OptionStore from "../common/OptionStore";
 import ToastStore from "../toast/ToastStore";
 import commonActions from "../common/commonActions";
