+++ src/panel/spoiler_card/SpoilerCardListContainer.js
@@ -2,14 +2,14 @@ import React from "react";
 import {observer} from "mobx-react";
 import {observable, computed} from "mobx";
 import {Button} from "@blueprintjs/core";
-import MainStore from "../MainStore";
+import MainStore from "../common/MainStore";
 import SpoilerCardList from "./SpoilerCardList";
 
 @observer
 class SpoilerCardListContainer extends React.Component {
 	@observable isShowingLess = true;
 	@observable SHOW_LESS_NUM = 5; // TODO: move this out in the global options store (where all defaults are set)
-	
+
 	@computed get numListsToShow() {
 		if (this.isShowingLess) {
 			return Math.min(this.SHOW_LESS_NUM, MainStore.spoilers.length);
@@ -20,7 +20,7 @@ class SpoilerCardListContainer extends React.Component {
 	}
 
 	@computed get SeeButton() {
-		if (MainStore.spoilers.length > this.numListsToShow && this.isShowingLess) { 
+		if (MainStore.spoilers.length > this.numListsToShow && this.isShowingLess) {
 			return <Button className="pt-intent-primary" text="See More" onClick={this.handleSeeMoreLess} />;
 		}
 		else if (!this.isShowingLess) {
@@ -45,7 +45,7 @@ class SpoilerCardListContainer extends React.Component {
 		return (
 			<div>
 				<SpoilerCardList spoilers={MainStore.spoilers.slice(0, this.numListsToShow)} />
-				{this.SeeButton}				
+				{this.SeeButton}
 			</div>
 		);
 	}
