+++ src/panel/title_bar/titleBarActions.js
@@ -1,5 +1,5 @@
 import {action} from "mobx";
-import MainStore from "../MainStore";
+import MainStore from "../common/MainStore";
 
 module.exports = {
 	showAddCard: action(showAddCard)
